+++ test/api/mock/MockIWebRequestTracer.h
@@ -48,6 +48,10 @@ namespace test
 
 		MOCK_METHOD(std::shared_ptr<openkit::IWebRequestTracer>, setBytesReceived, (int32_t), (override));
 
+		MOCK_METHOD(std::shared_ptr<openkit::IWebRequestTracer>, setBytesSent, (int64_t), (override));
+
+		MOCK_METHOD(std::shared_ptr<openkit::IWebRequestTracer>, setBytesReceived, (int64_t), (override));
+
 		MOCK_METHOD(std::shared_ptr<openkit::IWebRequestTracer>, start, (), (override));
 
 		MOCK_METHOD(void, stop, (int32_t), (override));